example/requirements.txt
MySQL-python
# django-socketio-chat (install using a .pth file in site-packages)
tornadio2
django-templatetag-handlebars
